Change 1 - Constraints removed from MOF model The OCL on the BusinessDiagram is removed because MOF does not support constraints.
Change 2 - messageBuildingBlock
The below structure was added in Copenhagen. This has two problems:
Two properties with the same name in one owner is illegal in UML. MessageDefinition can't have two properties both called messageBuildingBlock,
The ordering on the messageBuildingBlock isn't ordering across both ChoiceComponents and MessageComponents together, but it ordering over one and then the other.

There are multiple associations to Diagram subtypes for some types, so they have been given distinct property names. The main example is below:

Change 4 - Issues fixed in the eCore model
MessageDefinition.messageSchema was given an EType of SyntaxMessageScheme
MessageDefinition.marketPractice was given an EType of MarketPractice
BusinessComponentTrace.messageLayer was given an EType of MessageElement
MessageComponent.constraint was given an EType of MessageComponentRule

This is illegal because BusinessModel has two properties both named "mainRelevantConcept". This has been changed.