Following the 25th October 2007 teleplenum, the diagrams have been removed to a separate package and partly renamed as "representation".
Diagrams are also giving containment problems. The BusinessProcessCatalog diagrams will become items in a BusinessArea. This makes BusinessArea the primary sub-division of BusinessProcessCatalog. The DataDictionary diagrams will become items owned at the DataDictionary level. To replace the spaghetti of associations between the Diagrams and the things they might represent, this will be more concisely expressed in OCL. So for example a BusinessComponentDiagram is a kind of Diagram and is contained within the DataDictionary. The contents of a BusinessComponentDiagram are expressed in OCL. This will greatly simplify the metamodel.